features & fixes: - fix: add missing wtPos{modulator} value tree parameter - collect calculated modulators to one struct ModulatorCalculatedValues - wave table modulation feature(finally!) - frequency calculation encapsulated - morphing wave encapsulated

+++ Source/audio/wave/WaveTables.h
@@ -57,7 +57,7 @@ namespace audio {
             ++phaseOffset;
         }
 
-        float generateSample(double frequency, double sampleRate, int waveIndex, float phaseOffsetPercentage) {
+        float generateSample(double frequency, double sampleRate, size_t waveIndex, float phaseOffsetPercentage) {
             ZoneScoped;
             return waveTable[waveIndex]->generate((float)frequency, (float)sampleRate, phaseOffset, phaseOffsetPercentage);
         }
